+++ src/kinematics/CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 2.8)
 
-project(TestKinematics)
+#project(TestKinematics)
 
 find_package(Boost COMPONENTS system thread filesystem)
 find_package(Eigen3 REQUIRED)
@@ -16,8 +16,11 @@ link_directories(
 	${Boost_LIBRARY_DIR}
 )
 
-add_executable(TestKinematics Kinematics.cpp Link.cpp testKinematics.cpp Kinematics.h Link.h)
+#add_executable(TestKinematics Kinematics.cpp Link.cpp testKinematics.cpp Kinematics.h Link.h)
+add_executable(testCheckCoM testCheckCom.cpp Kinematics.cpp Jacobian.cpp Link.cpp)
 
-target_link_libraries(TestKinematics ${Boost_LIBRARIES})
+#target_link_libraries(TestKinematics ${Boost_LIBRARIES})
+target_link_libraries(testCheckCoM ${Boost_LIBRARIES})
 
-install(TARGETS TestKinematics RUNTIME DESTINATION bin)
+#nstall(TARGETS TestKinematics RUNTIME DESTINATION bin)
+install(TARGETS testCheckCoM RUNTIME DESTINATION bin)
